[update] strategy -> BinaryOperator
