--- src/main/java/domain/operation/ComputationStrategy.java
@@ -1,8 +1,0 @@
-package domain.operation;
-
-
-import domain.Operand;
-
-interface ComputationStrategy {
-    Operand compute(Operand input, Operand operand);
-}
