+++ src/test/java/domain/vo/OperationTest.java
@@ -1,7 +1,7 @@
 package domain.vo;
 
 import domain.Operand;
-import domain.operation.Operation;
+import domain.Operation;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -11,7 +11,7 @@ import org.junit.jupiter.params.provider.ValueSource;
 
 import java.util.stream.Stream;
 
-import static domain.operation.Operation.DIVISION;
+import static domain.Operation.DIVISION;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
